[rename] the components folders
