+++ src/scripts/components/logo/Logo.js
@@ -1,14 +1,3 @@
-/* import React from "react";
-import logo from "../../images/logo.svg";
-
-export default function Logo() {
-    return(
-        <div>
-            <img src={logo} alt="Logo"/>
-        </div>
-    )
-} */
-
 import React from "react";
 
 export default function Logo() {
