+++ src/scripts/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./components/app/App";
+import App from "./features/App/App";
 import "../scss/index.scss";
 
 ReactDOM.render(<App />, document.getElementById("root"));
