--- src/scripts/useForm.spec.js
@@ -1,14 +1,0 @@
-import { cleanup, render, screen } from "@testing-library/react";
-import useForm from "./useForm";
-
-describe("use form test", () => {
-
-    afterEach(cleanup);
-
-    //const { handleChange, values, handleSubmit, errors } = useForm();
-
-    it("", () => {
-
-    });
-
-})
